## Overview

Speeds up `get_string` in two ways:

1. Cache the `java/lang/String` class instead of calling `FindClass` repeatedly.
2. Use `IsInstanceOf` instead of `GetObjectClass` + `IsAssignableFrom`. `java/lang/String` is final, so these are equivalent.

Running the benchmark on my machine, change 1:

```
jni_get_string          time:   [445.86 ns 448.89 ns 454.23 ns]
                        change: [-53.617% -45.104% -39.388%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Changes 1+2:
```
jni_get_string          time:   [406.30 ns 406.37 ns 406.45 ns]
                        change: [-9.6565% -9.1695% -8.8758%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Also incidentally fixes #528.

**Questions/notes**

* I don't write Rust, so sorry if I made any basic errors.
* I tried to use `std::OnceLock`, but `get_or_try_init` is not yet stable, and I couldn't find a neat way to handle errors, hence pulling in `once_cell`. Open to ideas here.
* Let me know your thoughts on the changelog and doc changes.

Thanks!

### Definition of Done

- [x] There are no TODOs left in the code
- [x] Change is covered by [automated tests](https://github.com/jni-rs/jni-rs/blob/master/CONTRIBUTING.md#tests)
- [x] The [coding guidelines](https://github.com/jni-rs/jni-rs/blob/master/CONTRIBUTING.md#the-code-style) are followed
- [x] Public API has documentation
- [x] User-visible changes are mentioned in the Changelog
- [ ] The continuous integration build passes
  - TBD